Several layer types deprecated
ANNWhitenedLayer
ImageFromMemoryBuffer
KmeansLayer
LabelErrorLayer
LCALIFLayer
RunningAverageLayer
ShuffleLayer
