+++ src/CMakeLists.txt
@@ -47,6 +47,7 @@ set(SRC_SUBDIRS
   connections
   cudakernels
   delivery
+  deprecated
   include
   initv
   io
