+++ src/deprecated/LabelErrorLayer.cpp
@@ -1,13 +1,16 @@
 /*
- * ANNWhitenedLayer.cpp
+ * LabelErrorLayer.cpp
  *
- *  Created on: Feb 15, 2013
+ *  Created on: Nov 30, 2013
  *      Author: garkenyon
  */
 
-#include "ANNWhitenedLayer.hpp"
+// LabelErrorLayer was deprecated on Aug 15, 2018.
 
-void ANNWhitenedLayer_update_state(
+#include "LabelErrorLayer.hpp"
+#include "DeprecatedUpdateStateFunctions.h"
+
+void LabelErrorLayer_update_state(
       const int nbatch,
       const int numNeurons,
       const int nx,
@@ -24,31 +27,51 @@ void ANNWhitenedLayer_update_state(
       float *verticesA,
       float *slopes,
       float *GSynHead,
-      float *activity);
+      float *activity,
+      float errScale,
+      int isBinary);
 
 namespace PV {
 
-ANNWhitenedLayer::ANNWhitenedLayer() { initialize_base(); }
+LabelErrorLayer::LabelErrorLayer() { initialize_base(); }
 
-ANNWhitenedLayer::ANNWhitenedLayer(const char *name, HyPerCol *hc) {
+LabelErrorLayer::LabelErrorLayer(const char *name, HyPerCol *hc) {
    initialize_base();
    initialize(name, hc);
 }
 
-ANNWhitenedLayer::~ANNWhitenedLayer() {}
+LabelErrorLayer::~LabelErrorLayer() {}
 
-int ANNWhitenedLayer::initialize_base() {
-   numChannels = 3; // applyGSyn_ANNWhitenedLayer uses 3 channels
+int LabelErrorLayer::initialize_base() {
+   numChannels = 2;
+   errScale    = 1;
+   isBinary    = 1;
    return PV_SUCCESS;
 }
 
-int ANNWhitenedLayer::initialize(const char *name, HyPerCol *hc) {
-   ANNLayer::initialize(name, hc);
-   assert(numChannels == 3);
-   return PV_SUCCESS;
+int LabelErrorLayer::initialize(const char *name, HyPerCol *hc) {
+   WarnLog() << "LabelErrorLayer has been deprecated.\n";
+   int status = ANNLayer::initialize(name, hc);
+   assert(numChannels == 2);
+   return status;
 }
 
-Response::Status ANNWhitenedLayer::updateState(double time, double dt) {
+int LabelErrorLayer::ioParamsFillGroup(enum ParamsIOFlag ioFlag) {
+   int status = ANNLayer::ioParamsFillGroup(ioFlag);
+   ioParam_errScale(ioFlag);
+   ioParam_isBinary(ioFlag);
+   return status;
+}
+
+void LabelErrorLayer::ioParam_errScale(enum ParamsIOFlag ioFlag) {
+   parent->parameters()->ioParamValue(ioFlag, name, "errScale", &errScale, errScale);
+}
+
+void LabelErrorLayer::ioParam_isBinary(enum ParamsIOFlag ioFlag) {
+   parent->parameters()->ioParamValue(ioFlag, name, "isBinary", &isBinary, isBinary);
+}
+
+Response::Status LabelErrorLayer::updateState(double time, double dt) {
    const PVLayerLoc *loc = getLayerLoc();
    float *A              = clayer->activity->data;
    float *V              = getV();
@@ -59,7 +82,7 @@ Response::Status ANNWhitenedLayer::updateState(double time, double dt) {
    int nf                = loc->nf;
    int num_neurons       = nx * ny * nf;
    int nbatch            = loc->nbatch;
-   ANNWhitenedLayer_update_state(
+   LabelErrorLayer_update_state(
          nbatch,
          num_neurons,
          nx,
@@ -75,14 +98,18 @@ Response::Status ANNWhitenedLayer::updateState(double time, double dt) {
          verticesA,
          slopes,
          gSynHead,
-         A);
+         A,
+         errScale,
+         isBinary);
 
    return Response::SUCCESS;
 }
 
 } /* namespace PV */
 
-void ANNWhitenedLayer_update_state(
+// Kernel
+
+void LabelErrorLayer_update_state(
       const int nbatch,
       const int numNeurons,
       const int nx,
@@ -99,8 +126,10 @@ void ANNWhitenedLayer_update_state(
       float *verticesA,
       float *slopes,
       float *GSynHead,
-      float *activity) {
-   updateV_ANNWhitenedLayer(
+      float *activity,
+      const float errScale,
+      const int isBinary) {
+   updateV_LabelErrorLayer(
          nbatch,
          numNeurons,
          V,
@@ -116,5 +145,7 @@ void ANNWhitenedLayer_update_state(
          lt,
          rt,
          dn,
-         up);
+         up,
+         errScale,
+         isBinary);
 }
